Small changes to support later GHC versions
Why upgrade `req`?

The currect code uses a `Default` instance for `HttpConfig`,
which is only available in `req` >=1.0 && <= 1.2.1.
This effectively pins `retry` at 0.7.*,
which requires `ghc-prim` < 0.6, but later versions of `base` require
greater versions of `ghc-prim`.

Why relax `time` bounds?

Time required an old `TemplateHaskell`, which required an old `base`.

Why pin `aeson` to < 2?

Because this package doesn't suport `aeson` >= 2, and fails to build it
a greater version is chosen.